refactor(backend): extract TagTable and BlogTagsTable into separate file
