+++ backend/app/src/main/kotlin/my/backend/db/schema/BlogTable.kt
@@ -13,17 +13,3 @@ object BlogTable : Table("blogs") {
 
     override val primaryKey = PrimaryKey(id)
 }
-
-object TagTable : Table("tags") {
-    val id = integer("id").autoIncrement()
-    val name = varchar("name", 50).uniqueIndex()
-
-    override val primaryKey = PrimaryKey(id)
-}
-
-object BlogTagsTable : Table("blog_tags") {
-    val blogId = reference("blog_id", BlogTable.id)
-    val tagId = reference("tag_id", TagTable.id)
-
-    override val primaryKey = PrimaryKey(blogId, tagId)
-}
