+++ backend/app/src/main/kotlin/my/backend/db/schema/TagTable.kt
@@ -1,0 +1,17 @@
+package my.backend.db.schema
+
+import org.jetbrains.exposed.sql.Table
+
+object TagTable : Table("tags") {
+    val id = integer("id").autoIncrement()
+    val name = varchar("name", 50).uniqueIndex()
+
+    override val primaryKey = PrimaryKey(id)
+}
+
+object BlogTagsTable : Table("blog_tags") {
+    val blogId = reference("blog_id", BlogTable.id)
+    val tagId = reference("tag_id", TagTable.id)
+
+    override val primaryKey = PrimaryKey(blogId, tagId)
+}
